+++ core/src/Threads/Kokkos_Threads_Parallel.hpp
@@ -269,7 +269,7 @@ class ParallelFor<FunctorType, Kokkos::TeamPolicy<Properties...>,
 
   const FunctorType m_functor;
   const Policy m_policy;
-  const int m_shared;
+  const size_t m_shared;
 
   template <class TagType, class Schedule>
   inline static std::enable_if_t<std::is_void<TagType>::value &&
@@ -720,7 +720,7 @@ class ParallelReduce<FunctorType, Kokkos::TeamPolicy<Properties...>,
   const Policy m_policy;
   const ReducerType m_reducer;
   const pointer_type m_result_ptr;
-  const int m_shared;
+  const size_t m_shared;
 
   template <class TagType>
   inline static std::enable_if_t<std::is_void<TagType>::value> exec_team(
